Do first part of training on gcp
